+++ Python/missionWidget.py
@@ -49,7 +49,7 @@ class MissionWidget(QWidget):
         self.height = self.mainHeight+self.underBarHeight
         self.left = 0
         self.top = 0
-        self.setGeometry(self.left, self.top, self.width, self.height)
+        self.setGeometry(int(self.left), int(self.top), int(self.width), int(self.height))
         # Buffer to the right of labels for the vertical lines
         self.underBarWBuffer = 3 * self.gui.pixel_scale_ratio[0]
         self.mainBarWBuffer = 10 * self.gui.pixel_scale_ratio[0]
@@ -89,9 +89,9 @@ class MissionWidget(QWidget):
 
         self.CET_label.setText("CET-00:00:00")
         self.CET_label.setStyleSheet("color: white")
-        self.CET_label.setFixedHeight(self.mainHeight)
+        self.CET_label.setFixedHeight(int(self.mainHeight))
         self.CET_label.setAlignment(Qt.AlignLeft | Qt.AlignVCenter)
-        self.CET_label.move(0, 0-4*self.gui.pixel_scale_ratio[1])  # Nasty but makes it look more centered
+        self.CET_label.move(0, int(0-4*self.gui.pixel_scale_ratio[1]))  # Nasty but makes it look more centered
         self.CET_label.show()
         # The right position of the CET label, with a 10 pixel buffer
         self.CET_labelRPos = self.CET_label.width() + self.mainBarWBuffer
@@ -112,10 +112,10 @@ class MissionWidget(QWidget):
         self.dateTimeLabel.setFont(monospace_light_font)
         self.dateTimeLabel.setStyleSheet("color: white")
         self.dateTimeLabel.setText("November 25th 2020, 05:52pm")
-        self.dateTimeLabel.setFixedHeight(self.underBarHeight)
+        self.dateTimeLabel.setFixedHeight(int(self.underBarHeight))
         self.dateTimeLabel.setAlignment(Qt.AlignLeft | Qt.AlignVCenter)
         # self.dateTimeLabel.setStyleSheet("background-color: red")
-        self.dateTimeLabel.move(0, self.mainHeight-1)
+        self.dateTimeLabel.move(0, int(self.mainHeight-1))
         self.updateDateTimeLabel()
         self.dateTimeLabel.show()
         # The right position of the dateTimeLabel, with a  buffer
@@ -127,18 +127,18 @@ class MissionWidget(QWidget):
         self.titleLabel.setStyleSheet('QLabel {color: #ffcb05;}') # MASA Maize
         self.titleLabel.setFont(monospace_light_font)
         self.titleLabel.setText("")
-        self.titleLabel.setFixedHeight(self.underBarHeight)
+        self.titleLabel.setFixedHeight(int(self.underBarHeight))
         self.titleLabel.setAlignment(Qt.AlignLeft | Qt.AlignBottom)
-        self.titleLabel.move(self.dateTimeLabelRPos + self.underBarWBuffer, self.mainHeight-1)
+        self.titleLabel.move(int(self.dateTimeLabelRPos + self.underBarWBuffer), int(self.mainHeight-1))
         self.titleLabel.show()
 
         # Set all the indicators, move into position
         self.campaignIndicator = IndicatorLightWidget(self, self.gui, 'Campaign', 20, "Red", 14, 30, 5, 2)
         self.campaignIndicator.setToolTip("Campaign has not started")
-        self.campaignIndicator.move(self.CET_labelRPos, 0)
+        self.campaignIndicator.move(int(self.CET_labelRPos), 0)
 
         self.connectionIndicator = IndicatorLightWidget(self, self.gui, 'Connection', 20, "Red", 14, 30, 5, 2)
-        self.connectionIndicator.move(self.campaignIndicator.pos().x() + self.campaignIndicator.width(), 0)
+        self.connectionIndicator.move(int(self.campaignIndicator.pos().x() + self.campaignIndicator.width()), 0)
         self.connectionIndicator.setToolTip("No connection")
 
         self.commandIndicator = IndicatorLightWidget(self, self.gui, 'Command', 20, "Red", 14, 30, 5, 2)
@@ -158,10 +158,10 @@ class MissionWidget(QWidget):
         self.status_label.setFont(CET_font)
         self.status_label.setText("GUI Configuration")
         self.status_label.setStyleSheet("color: white")
-        self.status_label.setFixedHeight(self.mainHeight)
-        self.status_label.setFixedWidth(self.width - (self.systemIndicator.pos().x() + self.systemIndicator.width()))
+        self.status_label.setFixedHeight(int(self.mainHeight))
+        self.status_label.setFixedWidth(int(self.width - (self.systemIndicator.pos().x() + self.systemIndicator.width())))
         self.status_label.setAlignment(Qt.AlignLeft | Qt.AlignVCenter)
-        self.status_label.move(self.systemIndicator.pos().x() + self.systemIndicator.width() + self.mainBarWBuffer, 0 - 4 * self.gui.pixel_scale_ratio[1])  # Nasty but makes it look more centered
+        self.status_label.move(int(self.systemIndicator.pos().x() + self.systemIndicator.width() + self.mainBarWBuffer), int(0 - 4 * self.gui.pixel_scale_ratio[1]))  # Nasty but makes it look more centered
         self.status_label.show()
 
         # Connect the start of the run to function to allow updating
@@ -322,7 +322,7 @@ class MissionWidget(QWidget):
 
         # Border around the date time label
         path = QPainterPath()
-        pen.setWidth(Constants.line_width/2)
+        pen.setWidth(int(Constants.line_width/2))
         self.painter.setPen(pen)
         path.moveTo(0, self.height - 1)
         path.lineTo(self.dateTimeLabel.width() + self.underBarWBuffer, self.height - 1)
